Adjust TSV writer to exclude index
