+++ data_processor/data_processor/Data_Processor_PyQt6.py
@@ -84,11 +84,17 @@ except ImportError:
     PYARROW_AVAILABLE = False
 
 # Matplotlib imports for PyQt6
+import matplotlib
+matplotlib.use("QtAgg")
 from matplotlib.figure import Figure
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
+from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.backends.backend_qtagg import NavigationToolbar2QT as NavigationToolbar
 import matplotlib.dates as mdates
 import matplotlib.pyplot as plt
+
+from file_utils import FileFormatDetector, DataReader, DataWriter
+from threads import ProcessingThread, ConversionThread, FolderProcessingThread
+from folder_tool_tab import FolderToolTab
 
 # =============================================================================
 # WORKER FUNCTION FOR PARALLEL PROCESSING
@@ -263,154 +269,6 @@ class SplitConfig:
         """Convert MB to bytes."""
         return int(self.max_file_size_mb * 1024 * 1024)
 
-class FileFormatDetector:
-    """Detects file format based on extension and content."""
-    
-    @staticmethod
-    def detect_format(file_path: str) -> Optional[str]:
-        """Detect file format from extension and content."""
-        if not os.path.exists(file_path):
-            return None
-            
-        # Check extension first
-        ext = os.path.splitext(file_path)[1].lower()
-        extension_map = {
-            '.csv': 'csv',
-            '.tsv': 'tsv',
-            '.parquet': 'parquet',
-            '.xlsx': 'excel',
-            '.xls': 'excel',
-            '.json': 'json',
-            '.h5': 'hdf5',
-            '.hdf5': 'hdf5',
-            '.pkl': 'pickle',
-            '.npy': 'numpy',
-            '.mat': 'matlab',
-            '.feather': 'feather',
-            '.arrow': 'arrow',
-            '.db': 'sqlite'
-        }
-        
-        if ext in extension_map:
-            return extension_map[ext]
-        
-        # Try to detect from content
-        try:
-            with open(file_path, 'rb') as f:
-                header = f.read(1024)
-                
-            # Check for magic numbers
-            if header.startswith(b'PAR1'):
-                return 'parquet'
-            elif header.startswith(b'ARROW1'):
-                return 'arrow'
-            elif header.startswith(b'PK\x03\x04'):
-                return 'excel'  # ZIP file (Excel)
-            elif header.startswith(b'\x89PNG'):
-                return 'png'
-            elif header.startswith(b'SQLite format 3'):
-                return 'sqlite'
-                
-        except:
-            pass
-            
-        return None
-
-class DataReader:
-    """Handles reading data from various file formats."""
-    
-    @staticmethod
-    def read_file(file_path: str, format_type: str, **kwargs) -> pd.DataFrame:
-        """Read file based on format type."""
-        try:
-            if format_type == 'csv':
-                return pd.read_csv(file_path, **kwargs)
-            elif format_type == 'tsv':
-                return pd.read_csv(file_path, sep='\t', **kwargs)
-            elif format_type == 'parquet':
-                if PYARROW_AVAILABLE:
-                    return pq.read_table(file_path).to_pandas()
-                else:
-                    return pd.read_parquet(file_path, **kwargs)
-            elif format_type == 'excel':
-                return pd.read_excel(file_path, engine='openpyxl', **kwargs)
-            elif format_type == 'json':
-                return pd.read_json(file_path, **kwargs)
-            elif format_type == 'hdf5':
-                return pd.read_hdf(file_path, **kwargs)
-            elif format_type == 'pickle':
-                return pd.read_pickle(file_path)
-            elif format_type == 'numpy':
-                return pd.DataFrame(np.load(file_path))
-            elif format_type == 'matlab':
-                return pd.DataFrame(scipy.io.loadmat(file_path))
-            elif format_type == 'feather':
-                if PYARROW_AVAILABLE:
-                    return feather.read_feather(file_path)
-                else:
-                    return pd.read_feather(file_path)
-            elif format_type == 'arrow':
-                if PYARROW_AVAILABLE:
-                    return pa.ipc.open_file(file_path).read_pandas()
-                else:
-                    raise ImportError("PyArrow required for Arrow format")
-            elif format_type == 'sqlite':
-                return pd.read_sql_query("SELECT * FROM data", f"sqlite:///{file_path}")
-            else:
-                raise ValueError(f"Unsupported format: {format_type}")
-        except Exception as e:
-            raise Exception(f"Error reading {file_path}: {e}")
-
-class DataWriter:
-    """Handles writing data to various file formats."""
-    
-    @staticmethod
-    def write_file(df: pd.DataFrame, file_path: str, format_type: str, **kwargs) -> None:
-        """Write file based on format type."""
-        try:
-            if format_type == 'csv':
-                df.to_csv(file_path, index=True, **kwargs)
-            elif format_type == 'tsv':
-                df.to_csv(file_path, sep='\t', index=True, **kwargs)
-            elif format_type == 'parquet':
-                if PYARROW_AVAILABLE:
-                    table = pa.Table.from_pandas(df)
-                    pq.write_table(table, file_path, compression=kwargs.get('compression', 'snappy'))
-                else:
-                    df.to_parquet(file_path, **kwargs)
-            elif format_type == 'excel':
-                df.to_excel(file_path, index=True, engine='openpyxl', **kwargs)
-            elif format_type == 'json':
-                df.to_json(file_path, orient='records', **kwargs)
-            elif format_type == 'hdf5':
-                df.to_hdf(file_path, key='data', **kwargs)
-            elif format_type == 'pickle':
-                df.to_pickle(file_path)
-            elif format_type == 'numpy':
-                np.save(file_path, df.values)
-            elif format_type == 'matlab':
-                scipy.io.savemat(file_path, {'data': df.values, 'columns': df.columns.tolist()})
-            elif format_type == 'feather':
-                if PYARROW_AVAILABLE:
-                    feather.write_feather(df, file_path)
-                else:
-                    df.to_feather(file_path)
-            elif format_type == 'arrow':
-                if PYARROW_AVAILABLE:
-                    table = pa.Table.from_pandas(df)
-                    with pa.ipc.new_file(file_path, table.schema) as writer:
-                        writer.write_table(table)
-                else:
-                    raise ImportError("PyArrow required for Arrow format")
-            elif format_type == 'sqlite':
-                import sqlite3
-                conn = sqlite3.connect(file_path)
-                df.to_sql('data', conn, if_exists='replace', index=False)
-                conn.close()
-            else:
-                raise ValueError(f"Unsupported format: {format_type}")
-        except Exception as e:
-            raise Exception(f"Error writing {file_path}: {e}")
 
 # =============================================================================
 # MAIN APPLICATION CLASS
@@ -466,21 +324,23 @@ class CSVProcessorAppPyQt6(QMainWindow):
         # Create tab widget
         self.tab_widget = QTabWidget()
         main_layout.addWidget(self.tab_widget)
-        
+
+        # Status bar
+        self.status_bar = QStatusBar()
+        self.setStatusBar(self.status_bar)
+        self.status_bar.showMessage("Ready")
+
         # Create tabs
         self.create_processing_tab()
         self.create_plotting_tab()
         self.create_plots_list_tab()
         self.create_format_converter_tab()
-        self.create_folder_tool_tab()
+        self.folder_tab = FolderToolTab(self.status_bar)
+        self.tab_widget.addTab(self.folder_tab, "Folder Tool")
         self.create_dat_import_tab()
+
         self.create_help_tab()
-        
-        # Create status bar
-        self.status_bar = QStatusBar()
-        self.setStatusBar(self.status_bar)
-        self.status_bar.showMessage("Ready")
-        
+
         # Create menu bar
         self.create_menu_bar()
         
@@ -697,67 +557,6 @@ class CSVProcessorAppPyQt6(QMainWindow):
         
         self.tab_widget.addTab(tab, "Format Converter")
         
-    def create_folder_tool_tab(self):
-        """Create the folder tool tab."""
-        tab = QWidget()
-        layout = QVBoxLayout(tab)
-        
-        # Source folders
-        source_group = QGroupBox("Source Folders")
-        source_layout = QVBoxLayout(source_group)
-        
-        source_buttons_layout = QHBoxLayout()
-        self.folder_select_source_btn = QPushButton("Select Source Folders")
-        self.folder_select_source_btn.clicked.connect(self.folder_select_source_folders)
-        source_buttons_layout.addWidget(self.folder_select_source_btn)
-        
-        self.folder_clear_source_btn = QPushButton("Clear All")
-        self.folder_clear_source_btn.clicked.connect(self.folder_clear_source_folders)
-        source_buttons_layout.addWidget(self.folder_clear_source_btn)
-        
-        source_layout.addLayout(source_buttons_layout)
-        
-        self.folder_source_list = QListWidget()
-        source_layout.addWidget(self.folder_source_list)
-        
-        layout.addWidget(source_group)
-        
-        # Operation selection
-        operation_group = QGroupBox("Operation")
-        operation_layout = QVBoxLayout(operation_group)
-        
-        self.folder_operation_combo = QComboBox()
-        self.folder_operation_combo.addItems(['Combine', 'Flatten', 'Prune', 'Deduplicate', 'Analyze'])
-        self.folder_operation_combo.currentTextChanged.connect(self.on_folder_operation_changed)
-        operation_layout.addWidget(self.folder_operation_combo)
-        
-        layout.addWidget(operation_group)
-        
-        # Destination
-        dest_group = QGroupBox("Destination")
-        dest_layout = QHBoxLayout(dest_group)
-        
-        self.folder_dest_edit = QLineEdit()
-        self.folder_dest_edit.setPlaceholderText("Select destination folder...")
-        dest_layout.addWidget(self.folder_dest_edit)
-        
-        self.folder_select_dest_btn = QPushButton("Browse")
-        self.folder_select_dest_btn.clicked.connect(self.folder_select_dest_folder)
-        dest_layout.addWidget(self.folder_select_dest_btn)
-        
-        layout.addWidget(dest_group)
-        
-        # Run button
-        self.folder_run_btn = QPushButton("Run Operation")
-        self.folder_run_btn.clicked.connect(self.folder_run_processing)
-        layout.addWidget(self.folder_run_btn)
-        
-        # Progress
-        self.folder_progress = QProgressBar()
-        layout.addWidget(self.folder_progress)
-        
-        self.tab_widget.addTab(tab, "Folder Tool")
-        
     def create_dat_import_tab(self):
         """Create the DAT file import tab."""
         tab = QWidget()
@@ -939,7 +738,7 @@ class CSVProcessorAppPyQt6(QMainWindow):
         self.selected_signals = [item.text() for item in selected_items]
         
         # Start processing in a separate thread
-        self.processing_thread = ProcessingThread(self.data_files, self.settings, self.output_folder)
+        self.processing_thread = ProcessingThread(self.data_files, self.settings, self.output_folder, process_single_csv_file)
         self.processing_thread.progress_updated.connect(self.update_processing_progress)
         self.processing_thread.finished.connect(self.processing_finished)
         self.processing_thread.start()
@@ -1154,68 +953,6 @@ class CSVProcessorAppPyQt6(QMainWindow):
         QMessageBox.information(self, "Complete", "File conversion completed!")
         self.status_bar.showMessage("Conversion completed")
         
-    def folder_select_source_folders(self):
-        """Select source folders for folder operations."""
-        folders = QFileDialog.getExistingDirectories(self, "Select Source Folders")
-        if folders:
-            for folder in folders:
-                self.folder_source_list.addItem(folder)
-                
-    def folder_clear_source_folders(self):
-        """Clear all source folders."""
-        self.folder_source_list.clear()
-        
-    def on_folder_operation_changed(self, operation):
-        """Handle folder operation change."""
-        # Update UI based on selected operation
-        pass
-        
-    def folder_select_dest_folder(self):
-        """Select destination folder for folder operations."""
-        folder = QFileDialog.getExistingDirectory(self, "Select Destination Folder")
-        if folder:
-            self.folder_dest_edit.setText(folder)
-            
-    def folder_run_processing(self):
-        """Run the folder processing operation."""
-        if self.folder_source_list.count() == 0:
-            QMessageBox.warning(self, "Warning", "Please select source folders.")
-            return
-            
-        if not self.folder_dest_edit.text():
-            QMessageBox.warning(self, "Warning", "Please select destination folder.")
-            return
-            
-        operation = self.folder_operation_combo.currentText()
-        source_folders = self.get_folder_source_list()
-        dest_folder = self.folder_dest_edit.text()
-        
-        # Start folder processing in a separate thread
-        self.folder_thread = FolderProcessingThread(
-            source_folders,
-            dest_folder,
-            operation
-        )
-        self.folder_thread.progress_updated.connect(self.update_folder_progress)
-        self.folder_thread.finished.connect(self.folder_processing_finished)
-        self.folder_thread.start()
-        
-    def get_folder_source_list(self):
-        """Get list of source folders from widget."""
-        folders = []
-        for i in range(self.folder_source_list.count()):
-            folders.append(self.folder_source_list.item(i).text())
-        return folders
-        
-    def update_folder_progress(self, value):
-        """Update folder processing progress bar."""
-        self.folder_progress.setValue(value)
-        
-    def folder_processing_finished(self):
-        """Handle folder processing completion."""
-        QMessageBox.information(self, "Complete", "Folder processing completed!")
-        self.status_bar.showMessage("Folder processing completed")
-        
     def select_dat_file(self):
         """Select DAT file for import."""
         file_path, _ = QFileDialog.getOpenFileName(
@@ -1391,337 +1128,6 @@ class CSVProcessorAppPyQt6(QMainWindow):
 # WORKER THREADS
 # =============================================================================
 
-class ProcessingThread(QThread):
-    """Thread for processing files."""
-    progress_updated = pyqtSignal(int, str)
-    
-    def __init__(self, files, settings, output_folder):
-        super().__init__()
-        self.files = files
-        self.settings = settings
-        self.output_folder = output_folder
-        
-    def run(self):
-        """Run the processing."""
-        for i, file_path in enumerate(self.files):
-            self.progress_updated.emit(
-                int((i / len(self.files)) * 100),
-                f"Processing {os.path.basename(file_path)}..."
-            )
-            
-            # Process the file
-            result = process_single_csv_file(file_path, self.settings)
-            
-            if result is not None:
-                # Save the processed file
-                output_path = os.path.join(
-                    self.output_folder,
-                    f"processed_{os.path.basename(file_path)}"
-                )
-                result.to_csv(output_path)
-                
-        self.progress_updated.emit(100, "Processing completed")
-
-class ConversionThread(QThread):
-    """Thread for file conversion."""
-    progress_updated = pyqtSignal(int)
-    log_updated = pyqtSignal(str)
-    
-    def __init__(self, files, output_format, combine_files, output_dir):
-        super().__init__()
-        self.files = files
-        self.output_format = output_format
-        self.combine_files = combine_files
-        self.output_dir = output_dir
-        
-    def run(self):
-        """Run the conversion."""
-        try:
-            if self.combine_files:
-                self.convert_combined_files()
-            else:
-                self.convert_separate_files()
-        except Exception as e:
-            self.log_updated.emit(f"Error: {e}")
-            
-    def convert_combined_files(self):
-        """Convert all files into one combined file."""
-        self.log_updated.emit("Starting combined file conversion...")
-        
-        combined_data = []
-        for i, file_path in enumerate(self.files):
-            self.progress_updated.emit(int((i / len(self.files)) * 50))
-            self.log_updated.emit(f"Reading {os.path.basename(file_path)}...")
-            
-            try:
-                # Detect format and read file
-                format_type = FileFormatDetector.detect_format(file_path)
-                df = DataReader.read_file(file_path, format_type)
-                combined_data.append(df)
-            except Exception as e:
-                self.log_updated.emit(f"Error reading {file_path}: {e}")
-                
-        if combined_data:
-            # Combine all dataframes
-            combined_df = pd.concat(combined_data, ignore_index=True)
-            
-            # Save combined file
-            output_path = os.path.join(
-                self.output_dir,
-                f"combined_data.{self.output_format}"
-            )
-            
-            self.log_updated.emit(f"Saving combined file to {output_path}...")
-            DataWriter.write_file(combined_df, output_path, self.output_format)
-            self.log_updated.emit("Combined file conversion completed!")
-            
-        self.progress_updated.emit(100)
-        
-    def convert_separate_files(self):
-        """Convert each file separately."""
-        self.log_updated.emit("Starting separate file conversion...")
-        
-        for i, file_path in enumerate(self.files):
-            self.progress_updated.emit(int((i / len(self.files)) * 100))
-            self.log_updated.emit(f"Converting {os.path.basename(file_path)}...")
-            
-            try:
-                # Detect format and read file
-                format_type = FileFormatDetector.detect_format(file_path)
-                df = DataReader.read_file(file_path, format_type)
-                
-                # Generate output filename
-                base_name = os.path.splitext(os.path.basename(file_path))[0]
-                output_path = os.path.join(
-                    self.output_dir,
-                    f"{base_name}.{self.output_format}"
-                )
-                
-                # Write file
-                DataWriter.write_file(df, output_path, self.output_format)
-                self.log_updated.emit(f"Converted: {os.path.basename(output_path)}")
-                
-            except Exception as e:
-                self.log_updated.emit(f"Error converting {file_path}: {e}")
-                
-        self.log_updated.emit("Separate file conversion completed!")
-
-class FolderProcessingThread(QThread):
-    """Thread for folder processing operations."""
-    progress_updated = pyqtSignal(int)
-    
-    def __init__(self, source_folders, dest_folder, operation):
-        super().__init__()
-        self.source_folders = source_folders
-        self.dest_folder = dest_folder
-        self.operation = operation
-        
-    def run(self):
-        """Run the folder processing operation."""
-        try:
-            if self.operation == "Combine":
-                self.combine_operation()
-            elif self.operation == "Flatten":
-                self.flatten_operation()
-            elif self.operation == "Prune":
-                self.prune_operation()
-            elif self.operation == "Deduplicate":
-                self.deduplicate_operation()
-            elif self.operation == "Analyze":
-                self.analyze_operation()
-        except Exception as e:
-            print(f"Error in folder processing: {e}")
-            
-    def combine_operation(self):
-        """Combine files from multiple folders into one folder."""
-        self.progress_updated.emit(0)
-        
-        # Create destination folder if it doesn't exist
-        os.makedirs(self.dest_folder, exist_ok=True)
-        
-        file_count = 0
-        total_files = sum(len([f for f in os.listdir(folder) if os.path.isfile(os.path.join(folder, f))]) 
-                         for folder in self.source_folders)
-        
-        for folder in self.source_folders:
-            for file in os.listdir(folder):
-                file_path = os.path.join(folder, file)
-                if os.path.isfile(file_path):
-                    # Copy file to destination
-                    dest_path = os.path.join(self.dest_folder, file)
-                    if os.path.exists(dest_path):
-                        # Add suffix to avoid overwriting
-                        base, ext = os.path.splitext(file)
-                        counter = 1
-                        while os.path.exists(dest_path):
-                            dest_path = os.path.join(self.dest_folder, f"{base}_{counter}{ext}")
-                            counter += 1
-                    
-                    shutil.copy2(file_path, dest_path)
-                    file_count += 1
-                    self.progress_updated.emit(int((file_count / total_files) * 100))
-                    
-        self.progress_updated.emit(100)
-        
-    def flatten_operation(self):
-        """Flatten folder structure by moving all files to destination."""
-        self.progress_updated.emit(0)
-        
-        # Create destination folder if it doesn't exist
-        os.makedirs(self.dest_folder, exist_ok=True)
-        
-        file_count = 0
-        total_files = 0
-        
-        # Count total files first
-        for folder in self.source_folders:
-            for root, dirs, files in os.walk(folder):
-                total_files += len(files)
-        
-        for folder in self.source_folders:
-            for root, dirs, files in os.walk(folder):
-                for file in files:
-                    file_path = os.path.join(root, file)
-                    # Copy file to destination
-                    dest_path = os.path.join(self.dest_folder, file)
-                    if os.path.exists(dest_path):
-                        # Add suffix to avoid overwriting
-                        base, ext = os.path.splitext(file)
-                        counter = 1
-                        while os.path.exists(dest_path):
-                            dest_path = os.path.join(self.dest_folder, f"{base}_{counter}{ext}")
-                            counter += 1
-                    
-                    shutil.copy2(file_path, dest_path)
-                    file_count += 1
-                    self.progress_updated.emit(int((file_count / total_files) * 100))
-                    
-        self.progress_updated.emit(100)
-        
-    def prune_operation(self):
-        """Remove empty folders and duplicate files."""
-        self.progress_updated.emit(0)
-        
-        # Create destination folder if it doesn't exist
-        os.makedirs(self.dest_folder, exist_ok=True)
-        
-        file_count = 0
-        total_files = sum(len([f for f in os.listdir(folder) if os.path.isfile(os.path.join(folder, f))]) 
-                         for folder in self.source_folders)
-        
-        seen_files = set()
-        
-        for folder in self.source_folders:
-            for file in os.listdir(folder):
-                file_path = os.path.join(folder, file)
-                if os.path.isfile(file_path):
-                    # Check if file is duplicate (simple size-based check)
-                    file_size = os.path.getsize(file_path)
-                    file_key = (file, file_size)
-                    
-                    if file_key not in seen_files:
-                        seen_files.add(file_key)
-                        # Copy file to destination
-                        dest_path = os.path.join(self.dest_folder, file)
-                        shutil.copy2(file_path, dest_path)
-                    
-                    file_count += 1
-                    self.progress_updated.emit(int((file_count / total_files) * 100))
-                    
-        self.progress_updated.emit(100)
-        
-    def deduplicate_operation(self):
-        """Remove duplicate files based on content."""
-        self.progress_updated.emit(0)
-        
-        # Create destination folder if it doesn't exist
-        os.makedirs(self.dest_folder, exist_ok=True)
-        
-        file_count = 0
-        total_files = sum(len([f for f in os.listdir(folder) if os.path.isfile(os.path.join(folder, f))]) 
-                         for folder in self.source_folders)
-        
-        seen_hashes = set()
-        
-        for folder in self.source_folders:
-            for file in os.listdir(folder):
-                file_path = os.path.join(folder, file)
-                if os.path.isfile(file_path):
-                    try:
-                        # Calculate file hash (simple MD5)
-                        import hashlib
-                        with open(file_path, 'rb') as f:
-                            file_hash = hashlib.md5(f.read()).hexdigest()
-                        
-                        if file_hash not in seen_hashes:
-                            seen_hashes.add(file_hash)
-                            # Copy file to destination
-                            dest_path = os.path.join(self.dest_folder, file)
-                            shutil.copy2(file_path, dest_path)
-                    except:
-                        pass
-                    
-                    file_count += 1
-                    self.progress_updated.emit(int((file_count / total_files) * 100))
-                    
-        self.progress_updated.emit(100)
-        
-    def analyze_operation(self):
-        """Analyze folder contents and generate report."""
-        self.progress_updated.emit(0)
-        
-        # Create destination folder if it doesn't exist
-        os.makedirs(self.dest_folder, exist_ok=True)
-        
-        analysis_results = []
-        folder_count = 0
-        total_folders = len(self.source_folders)
-        
-        for folder in self.source_folders:
-            folder_stats = {
-                'folder': folder,
-                'total_files': 0,
-                'total_size': 0,
-                'file_types': {},
-                'subfolders': 0
-            }
-            
-            for root, dirs, files in os.walk(folder):
-                folder_stats['subfolders'] += len(dirs)
-                for file in files:
-                    file_path = os.path.join(root, file)
-                    try:
-                        file_size = os.path.getsize(file_path)
-                        folder_stats['total_files'] += 1
-                        folder_stats['total_size'] += file_size
-                        
-                        # Count file types
-                        ext = os.path.splitext(file)[1].lower()
-                        folder_stats['file_types'][ext] = folder_stats['file_types'].get(ext, 0) + 1
-                    except:
-                        pass
-            
-            analysis_results.append(folder_stats)
-            folder_count += 1
-            self.progress_updated.emit(int((folder_count / total_folders) * 100))
-        
-        # Save analysis report
-        report_path = os.path.join(self.dest_folder, "folder_analysis_report.txt")
-        with open(report_path, 'w') as f:
-            f.write("Folder Analysis Report\n")
-            f.write("=" * 50 + "\n\n")
-            
-            for stats in analysis_results:
-                f.write(f"Folder: {stats['folder']}\n")
-                f.write(f"Total Files: {stats['total_files']}\n")
-                f.write(f"Total Size: {stats['total_size'] / (1024*1024):.2f} MB\n")
-                f.write(f"Subfolders: {stats['subfolders']}\n")
-                f.write("File Types:\n")
-                for ext, count in stats['file_types'].items():
-                    f.write(f"  {ext}: {count}\n")
-                f.write("\n")
-        
-        self.progress_updated.emit(100)
 
 # =============================================================================
 # MAIN ENTRY POINT
